courseAppv1/requirements.txt
anyio==4.4.0
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.4.6
decorator==4.4.2
distlib==0.3.9
Django==5.1.2
django-ckeditor==6.7.1
django-js-asset==2.2.0
filelock==3.16.1
git-filter-repo==2.45.0
gTTS==2.5.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
imageio==2.35.1
imageio-ffmpeg==0.5.1
moviepy==1.0.3
mysqlclient==2.2.5
numpy==2.1.1
pillow==11.0.0
platformdirs==4.3.6
playsound==1.3.0
proglog==0.1.10
PyAudio==0.2.14
pygame==2.6.0
PyMySQL==1.1.1
pynput==1.7.7
pypiwin32==223
pyttsx3==2.91
pytube==15.0.0
pywin32==306
requests==2.32.3
setuptools==60.2.0
six==1.16.0
sniffio==1.3.1
SpeechRecognition==3.10.4
sqlparse==0.5.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.2
virtualenv==20.27.0
wheel==0.37.1
youtube-dl==2021.12.17
youtube-search-python==1.6.6
